% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuna_lasso.R
\name{repeat_lasso}
\alias{repeat_lasso}
\title{repeat_lasso}
\usage{
repeat_lasso(trials, my_x, my_y, my_alpha = 1, my_family = "gaussian")
}
\arguments{
\item{trials}{number of times to repeat cross-validation on the lasso}

\item{my_x}{matrix of numeric predictors}

\item{my_y}{matrix (1D) response}

\item{my_alpha}{numeric from 0 (ridge) to 1 ()}

\item{my_family}{string "gaussian", 'binomial"}
}
\value{
list, first element is data frame summarizing fits, second element contains all of the glmnet cvfit objects
}
\description{
call run_lasso over a specified number of trials to see the frequency with which the regularization
proceedure include each coefficient
}
\examples{
repeat_lasso(10,
my_x = as.matrix(mtcars[1:15,2:dim(mtcars)[2]]),
my_y = as.matrix(mtcars[1:15,1]))

}
