% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuna_lasso.R
\name{run_lasso}
\alias{run_lasso}
\title{run_lasso}
\usage{
run_lasso(x, y, s, alpha, family)
}
\arguments{
\item{x}{matrix of numeric predictors}

\item{y}{matrix (1D) response}

\item{s}{int randomization seed}

\item{alpha}{numeric from 0 (ridge) to 1 ()}

\item{family}{string "gaussian", 'binomial", "poisson'}
}
\value{
list of two elements (1) cvfit , cv.glmnet full output and (2) r data.frame of non zero coefficients
}
\description{
wrapper around cv.glmnet() to run lasso regression with a specific seed
}
