% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuna_plots.R
\name{visualize_all_cverror}
\alias{visualize_all_cverror}
\title{visualize_all_cverror}
\usage{
visualize_all_cverror(x)
}
\arguments{
\item{x}{list generated by the function tuna::repeat_lasso}
}
\value{
plots in r base graphics
}
\description{
shows all CV error for all lambdas across all intialization conditoins
}
