% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuna_plots.R
\name{visualize_coef_density}
\alias{visualize_coef_density}
\title{visualize_coef_density}
\usage{
visualize_coef_density(x)
}
\arguments{
\item{x}{list generated by the function tuna::repeat_lasso}
}
\value{
ggplot2 object
}
\description{
density curves for raw coefficient values
}
